+++ tests/test_files/limma_tests/case3/expected_limma_script_3.R
@@ -4,13 +4,16 @@ design_matrix <- read.table("tests/test_files/test_design_matrix.csv", header=TR
 condition <- factor(design_matrix$condition, levels=c("cond1", "cond2", "cond3"))
 replicate <- factor(design_matrix$replicate, levels=c("rep1", "rep2", "rep3"))
 
-design <- model.matrix(~ condition + replicate)
+design <- model.matrix(~ condition)
 colnames(design)[1] <- "Intercept"
 
 count_data <- read.table("tests/test_files/big_counted.csv", header=TRUE, sep= ",", row.names = 1)
 voom_object <- voom(count_data, design, plot=FALSE, save.plot=TRUE)
 
-fit <- lmFit(voom_object, design)
+cor <- duplicateCorrelation(voom_object, design, block=replicate)
+if (cor$consensus.correlation > 0) { #only include random effect if the correlation is positive
+  fit <- lmFit(voom_object, design, block=replicate, correlation=cor$consensus.correlation)
+}else {   fit <- lmFit(voom_object, design)}
 
 contrast <- makeContrasts("conditioncond2", levels = design)
 contrast_fit <- contrasts.fit(fit, contrast)